--- packages/components/editor-toolbar/EditorToolbar.mdx
@@ -1,11 +1,0 @@
----
-title: 'EditorToolbar'
-type: 'component'
-status: 'stable'
-slug: /components/editor-toolbar/
-github: 'https://github.com/contentful/forma-36/tree/master/packages/components/editor-toolbar'
-storybook: 'https://f36-storybook.contentful.com/?path=/story/components-editor-toolbar--default'
-typescript: ./src/EditorToolbar.tsx
----
-
-# EditorToolbar
