--- packages/components/editor-toolbar/src/EditorToolbar.styles.ts
@@ -1,8 +1,0 @@
-import { css } from 'emotion';
-import tokens from '@contentful/f36-tokens';
-
-export const styles = {
-  editorToolbar: css({
-    color: tokens.colorRedBase,
-  }),
-};
