+++ packages/components/editor-toolbar/src/index.ts
@@ -1,2 +1,6 @@
 export { EditorToolbar } from './EditorToolbar';
 export type { EditorToolbarProps } from './EditorToolbar';
+export { EditorToolbarButton } from './EditorToolbarButton/EditorToolbarButton';
+export type { EditorToolbarButtonProps } from './EditorToolbarButton/EditorToolbarButton';
+export { EditorToolbarDivider } from './EditorToolbarDivider/EditorToolbarDivider';
+export type { EditorToolbarDividerProps } from './EditorToolbarDivider/EditorToolbarDivider';
