--- packages/forma-36-react-components/src/components/EditorToolbar/index.ts
@@ -1,6 +1,0 @@
-export { EditorToolbar } from './EditorToolbar';
-export type { EditorToolbarProps } from './EditorToolbar';
-export { EditorToolbarButton } from './EditorToolbarButton/EditorToolbarButton';
-export type { EditorToolbarButtonProps } from './EditorToolbarButton/EditorToolbarButton';
-export { EditorToolbarDivider } from './EditorToolbarDivider/EditorToolbarDivider';
-export type { EditorToolbarDividerProps } from './EditorToolbarDivider/EditorToolbarDivider';
